Added error codes and a standart error handling.
- Added ErrorCode struct.
- Added code and message for each error in the API.
- Added Error responses for sensors controllers.
- Added Error responses for groups controllers.
- Added Error responses for nodes controllers.

+++ internal/api/controllers/nodes.go
@@ -14,7 +14,6 @@ func GetNodes(w http.ResponseWriter, r *http.Request) {
 	logger := middlewares.ContextWithLogger(r.Context())
 	nodes := db.NodesData.GetNodes()
 
-	// TODO(#19) - Improve error handling
 	err := responses.SendJSONResponse(w, http.StatusOK, responses.ItemsResponse{
 		Code:    0,
 		Message: "All selected nodes information was returned",
@@ -22,22 +21,23 @@ func GetNodes(w http.ResponseWriter, r *http.Request) {
 	})
 	if err != nil {
 		logger.Error("Error sending response", "error", err.Error())
+		responses.SendErrorResponse(w, http.StatusInternalServerError, responses.InternalErrorCodes.ResponseError)
 		return
 	}
 }
 
 func GetNodeById(w http.ResponseWriter, r *http.Request) {
 	logger := middlewares.ContextWithLogger(r.Context())
+
 	// TODO(#20) - Validate Query Params
 	id := mux.Vars(r)["id"]
-	// TODO(#19) - Improve error handling
 	node, err := db.NodesData.GetNodeById(id)
 	if err != nil {
 		logger.Error("Error searching for node", "node_id", id, "error", err.Error())
+		responses.SendErrorResponse(w, http.StatusNotFound, responses.NodeErrorCodes.NotFound)
 		return
 	}
 
-	// TODO(#19) - Improve error handling
 	err = responses.SendJSONResponse(w, http.StatusOK, responses.ItemsResponse{
 		Code:    0,
 		Message: "All selected node information was returned",
@@ -45,17 +45,19 @@ func GetNodeById(w http.ResponseWriter, r *http.Request) {
 	})
 	if err != nil {
 		logger.Error("Error sending response", "error", err.Error())
+		responses.SendErrorResponse(w, http.StatusInternalServerError, responses.InternalErrorCodes.ResponseError)
 		return
 	}
 }
 
 func CreateNode(w http.ResponseWriter, r *http.Request) {
 	logger := middlewares.ContextWithLogger(r.Context())
+
 	var newNode *models.Node
-	// TODO(#19) - Improve error handling
 	err := json.NewDecoder(r.Body).Decode(&newNode)
 	if err != nil {
 		logger.Error("Error decoding new node", "body", r.Body, "error", err.Error())
+		responses.SendErrorResponse(w, http.StatusBadRequest, responses.NodeErrorCodes.InvalidSchema)
 		return
 	}
 
@@ -67,20 +69,27 @@ func CreateNode(w http.ResponseWriter, r *http.Request) {
 	})
 	if err != nil {
 		logger.Error("Error sending response", "error", err.Error())
+		responses.SendErrorResponse(w, http.StatusInternalServerError, responses.InternalErrorCodes.ResponseError)
 		return
 	}
 }
 
 func UpdateNode(w http.ResponseWriter, r *http.Request) {
 	logger := middlewares.ContextWithLogger(r.Context())
+
 	// TODO(#20) - Validate Query Params
 	id := mux.Vars(r)["id"]
+	if !db.GroupsData.IdExists(id) {
+		logger.Error("Error updating group", "node_id", id, "error", "group doesnt exist")
+		responses.SendErrorResponse(w, http.StatusNotFound, responses.NodeErrorCodes.NotFound)
+		return
+	}
 
 	var node *models.Node
-	// TODO(#19) - Improve error handling
 	err := json.NewDecoder(r.Body).Decode(&node)
 	if err != nil {
 		logger.Error("Error decoding new node", "body", r.Body, "error", err.Error())
+		responses.SendErrorResponse(w, http.StatusBadRequest, responses.NodeErrorCodes.InvalidSchema)
 		return
 	}
 	node.ID = id
@@ -88,10 +97,10 @@ func UpdateNode(w http.ResponseWriter, r *http.Request) {
 	node, err = db.NodesData.UpdateNode(node)
 	if err != nil {
 		logger.Error("Error updating nodes", "node_id", id, "error", err.Error())
+		responses.SendErrorResponse(w, http.StatusInternalServerError, responses.NodeErrorCodes.UpdateFailed)
 		return
 	}
 
-	// TODO(#19) - Improve error handling
 	err = responses.SendJSONResponse(w, http.StatusOK, responses.ItemsResponse{
 		Code:    0,
 		Message: "The selected node was updated",
@@ -99,6 +108,7 @@ func UpdateNode(w http.ResponseWriter, r *http.Request) {
 	})
 	if err != nil {
 		logger.Error("Error sending response", "error", err.Error())
+		responses.SendErrorResponse(w, http.StatusInternalServerError, responses.InternalErrorCodes.ResponseError)
 		return
 	}
 }
@@ -107,11 +117,16 @@ func DeleteNode(w http.ResponseWriter, r *http.Request) {
 	logger := middlewares.ContextWithLogger(r.Context())
 	// TODO(#20) - Validate Query Params
 	id := mux.Vars(r)["id"]
-	err := db.NodesData.DeleteNode(id)
+	if !db.GroupsData.IdExists(id) {
+		logger.Error("Error deleting group", "node_id", id, "error", "group doesnt exist")
+		responses.SendErrorResponse(w, http.StatusNotFound, responses.NodeErrorCodes.NotFound)
+		return
+	}
 
-	// TODO(#19) - Improve error handling
+	err := db.NodesData.DeleteNode(id)
 	if err != nil {
 		logger.Error("Error deleting group", "group_id", id, "error", err.Error())
+		responses.SendErrorResponse(w, http.StatusInternalServerError, responses.NodeErrorCodes.DeleteFailed)
 		return
 	}
 
@@ -121,6 +136,7 @@ func DeleteNode(w http.ResponseWriter, r *http.Request) {
 	})
 	if err != nil {
 		logger.Error("Error sending response", "error", err.Error())
+		responses.SendErrorResponse(w, http.StatusInternalServerError, responses.InternalErrorCodes.ResponseError)
 		return
 	}
 }
